+++ selene-lib/src/lints.rs
@@ -210,7 +210,7 @@ impl Label {
 #[derive(Clone, Debug)]
 pub struct Context {
     pub standard_library: StandardLibrary,
-    pub standard_library_is_set: bool,
+    pub user_set_standard_library: Option<Vec<String>>,
 }
 
 impl Context {