+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ## [Unreleased](https://github.com/Kampfkarren/selene/compare/0.23.0...HEAD)
 ### Added
 - Added new [`manual_table_clone` lint](https://kampfkarren.github.io/selene/lints/manual_table_clone.html), which will catch manual re-implementations of `table.clone` in Luau.
+- Added `filename` field to diagnostic message labels in JSON output, indicating for which file the message was generated (#453)
 
 ### Changed
 - Improved the error message for using a standard library that can be detected as outdated.